refactor(decision): rename getManualSelectionState to listSelectionCandidates
Drops the state gating from the list endpoint — `selectionsConfirmed` already lives on `getInstance`, so the list endpoint just returns `{ proposals }`. The name drops "Manual" so the endpoint can also be used to preview candidates before a phase advance. State-resolution logic moves to its own file (`resolveManualSelectionStatus.ts`), with its `instance` parameter now required and typed (both callers already supplied it).

Also removes the service-level `proposalIds.length === 0` check in submitManualSelection — the router's Zod `.min(1)` already enforces it.
