+++ packages/common/src/services/decision/index.ts
@@ -44,7 +44,8 @@ export * from './acceptProposalInvite';
 // Proposal management
 export * from './proposalDataSchema';
 export * from './isLegacyInstance';
-export * from './getManualSelectionState';
+export * from './listSelectionCandidates';
+export * from './resolveManualSelectionStatus';
 export * from './submitManualSelection';
 export * from './getProposalsForPhase';
 export * from './createProposal';
